-[DM][LOCATIONSERVICE] Deactivating all the deprecated constructors. There is a lot of code that depends on the location and that is obsolete. -[ACCEPTANCE] Updated the mining tests to the new EsiLocation.

+++ NeoCom.DataManagement/src/main/java-adapter/org/dimensinfin/eveonline/neocom/adapters/ESIDataAdapter.java
@@ -176,7 +176,7 @@ public class ESIDataAdapter {
 		return Futures.immediateFuture(new MarketDataSet(itemId, side));
 	}
 
-	public EsiLocation searchLocation4Id( final Long locationId ) {
+	public EsiLocation searchLocation4Id( final Integer locationId ) {
 		return this.locationCatalogService.searchLocation4Id(locationId);
 	}
